WebRtcIsac_UpdateBwEstimate et al.: Type byte streams as uint8, not uint16
This patch changes the signature of WebRtcIsac_UpdateBwEstimate,
WebRtcIsacfix_UpdateBwEstimate, and WebRtcIsacfix_UpdateBwEstimate1 so
that they expect the encoded data to be uint8 arrays, not uint16,
which is more natural. The implementations of the functions are left
unchanged for now.

BUG=909
R=aluebs@webrtc.org, bjornv@webrtc.org, henrik.lundin@webrtc.org, turaj@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/25729004

git-svn-id: http://webrtc.googlecode.com/svn/trunk@7430 4adac7df-926f-26a2-2b94-8c16560cd09d

+++ webrtc/modules/audio_coding/codecs/isac/fix/interface/isacfix.h
@@ -201,7 +201,7 @@ extern "C" {
    */
 
   int16_t WebRtcIsacfix_UpdateBwEstimate1(ISACFIX_MainStruct *ISAC_main_inst,
-                                          const uint16_t *encoded,
+                                          const uint8_t* encoded,
                                           int32_t  packet_size,
                                           uint16_t rtp_seq_number,
                                           uint32_t arr_ts);
@@ -226,7 +226,7 @@ extern "C" {
    */
 
   int16_t WebRtcIsacfix_UpdateBwEstimate(ISACFIX_MainStruct *ISAC_main_inst,
-                                         const uint16_t   *encoded,
+                                         const uint8_t* encoded,
                                          int32_t          packet_size,
                                          uint16_t         rtp_seq_number,
                                          uint32_t         send_ts,
